3rdparty/portaudio: Updated to latest upstream version. (#11604)
Up-to-date with revision 24c8d575e588d557d28f4011becb753421346860.  Resolves issues building with Visual Studio.

Enabled PortAudio when building with Visual Studio and clang-cl.

docs: Removed note about duplicate GUID symbols in PortAudio when built with MSVC.

+++ 3rdparty/portaudio/qa/paqa_latency.c
@@ -61,8 +61,8 @@ typedef struct
     int left_phase;
     int right_phase;
     char message[20];
-    int minFramesPerBuffer;
-    int maxFramesPerBuffer;
+    unsigned long minFramesPerBuffer;
+    unsigned long maxFramesPerBuffer;
     int callbackCount;
     PaTime minDeltaDacTime;
     PaTime maxDeltaDacTime;
@@ -170,8 +170,8 @@ PaError paqaCheckLatency( PaStreamParameters *outputParamsPtr,
     printf("Play for %d seconds.\n", NUM_SECONDS );
     Pa_Sleep( NUM_SECONDS * 1000 );
 
-    printf("  minFramesPerBuffer = %4d\n", dataPtr->minFramesPerBuffer );
-    printf("  maxFramesPerBuffer = %4d\n", dataPtr->maxFramesPerBuffer );
+    printf("  minFramesPerBuffer = %4lu\n", dataPtr->minFramesPerBuffer );
+    printf("  maxFramesPerBuffer = %4lu\n", dataPtr->maxFramesPerBuffer );
     printf("  minDeltaDacTime = %f\n", dataPtr->minDeltaDacTime );
     printf("  maxDeltaDacTime = %f\n", dataPtr->maxDeltaDacTime );
